+++ python/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `Name.is_valid()` now catches additional errors involving extra spaces between characters.
 
+## Removed
+
+- Removed support for setting `year` on the Paper level; can only be set on Volume now.
+
 ## [1.3.0] — 2026-07-31
 
 ### Added